[#3] Addition dependencies information.
I'm not sure what root is needed for at the begining, polkit isn't
used later for changing /etc/hosts for NFS; but it is required, and
unless a gui app/helper is running, it needs to be piped to the tty
agent.

NFS is a given it's just not mentioned, and usually at least
firewalled off.

Also adds note that from the commandline you might need to specify
the provider. The alternative may be to mention the method with the
virtualbox GUI?