Remove service provider access to content overview page
This conflicts with how killfile (trash bin functionality) works
because node that han be sent to trash still appear on the content
listing page, but it is not possible to perform any operation on
them and trying to visit them yields an access denied page. All this
is because the role does not (and should not) have any killfile
permission.