Update related programs and events view
Note that the views_distinct setting is not being exported.
To configure it, go to the view and for any disply update the
Content: Nid setting in the fields section. For 'filter/aggregate
this field' in the 'views distinct setting' select the option:
'Filter Repeats' and apply it to all the displays.